# app/static/swagger.yaml
...
swagger: '2.0'
info:
  title: "Service Ticket Management System API"
  description: "API to manage a Mechanic Service Ticket System"
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"

# =========================
# Security Definitions
# =========================
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header

# =========================
# Paths
# =========================
paths: 
...
  /mechanics/login: #Mechanics Login
    post: 
      tags:
        - Mechanics
      summary: Mechanic Login (generate token)
      description: This endpoint allows a mechanic to log in.
      parameters:
        - in: body
          name: body
          description: Login credentials
          required: true
          schema:
            $ref: '#/definitions/LoginPayload'
      responses:
        200:
          description: Login successful
          schema:
            $ref: '#/definitions/LoginResponse'
          examples:
            application/json:
              {
                "status": "success",
                "message": "Successfully logged in.",
                "token": "eyJhbGciOiJIUz....."
              }
        401:
          description: Invalid credentials
          schema:
            $ref: '#/definitions/MessageSchemaResponse'
          examples:
            application/json:
              message: Invalid email or password.
              status: error
  
  # --------- Mechanics CRUD ---------
  /mechanics: #Mechanics CRUD
    post: #Create
      tags:
        - Mechanics
      summary: Create a new mechanic
      description: This endpoint allows you to create a new mechanic.
      parameters:
        - in: body
          name: body
          description: Mechanic object that needs to be added
          required: true
          schema:
            $ref: '#/definitions/MechanicPayload'
      responses:
        201:
          description: Mechanic created successfully
          schema:
            $ref: '#/definitions/MechanicResponse'
          examples:
            application/json:
              {
                "id": 1,
                "name": "John Doe",
                "phone": "1234567890",
                "email": "john@app.com",
                "salary": 50000
              }
        400:
          description: Email already exists
          schema:
            $ref: '#/definitions/MessageSchemaResponse'
          examples:
            application/json:
              message: Email already exists.
              stastus: error     
    get: #Read
      tags:
        - Mechanics
      summary: Get mechanics (with pagination)
      description: This endpoint returns a list of mechanics with pagination. 
        **You can use the `page` and `per_page` query parameters to control pagination.**
      parameters:
        - in: query
          name: page
          type: integer
          description: Page number for pagination.
        - in: query
          name: per_page
          type: integer
          description: Number of mechanics per page.
      responses:
        200:
          description: Return all mechanics
          schema:
            $ref: '#/definitions/MechanicListResponse'
          examples:
            application/json:
              [
                {
                  "id": 1,
                  "name": "John Doe",
                  "phone": "1234567890",
                  "email": "john@app.com",
                  "salary": 50000
                }
              ]
    put: #Update
      tags:
        - Mechanics
      summary: Update a mechanic (Token required)
      description: Update an existing mechanic's details.
      security:
        - bearerAuth: []
      parameters:
        - in: body
          name: body
          description: Information used to update a mechanic
          required: true
          schema:
            $ref: '#/definitions/MechanicPayload'
      responses:
        200:
          description: Mechanic updated successfully
          schema:
            $ref: '#/definitions/MechanicResponse'
          examples:
            application/json:
              {
                "id": 1,
                "name": "John Doe",
                "phone": "1234567890",
                "email": "john@app.com",
                "salary": 50000,
              }
        400:
          description: Email already exists
          schema:
            $ref: '#/definitions/MessageSchemaResponse'
          examples:
            application/json:
              message: Email already exists.
              status: error          
    delete: #Delete
      tags:
        - Mechanics
      summary: Delete a mechanic (Token required)
      description: This endpoint allows you to delete a mechanic.
      security:
        - bearerAuth: []
      responses:
        200:
          description: Mechanic deleted successfully
          schema:
            $ref: '#/definitions/MessageSchemaResponse'
          examples:
            application/json:
              message: Successfully deleted mechanic 1.
              status: success

  # --------- Mechanics Get By ID ---------            
  /mechanics/{id}:
    get: #Read
      tags:
        - Mechanics
      summary: Get mechanic by ID
      description: Retrieve a mechanic's details by their unique ID.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            minimum: 1
            example: 1
          description: The ID of the mechanic to retrieve.
      responses:
        200:
          description: Mechanic found
          schema:
            $ref: '#/definitions/MechanicResponse'
          examples:
            application/json:
              {
                "id": 1,
                "name": "John Doe",
                "phone": "1234567890",
                "email": "john@app.com",
                "salary": 50000,
                "service_tickets": []
              }
        404:
          description: Mechanic not found
          schema:
            $ref: '#/definitions/MessageSchemaResponse'
          examples:
            application/json:
              message: Invalid mechanic
              status: error
              
  # --------- Mechanics Popular ---------
...
  /mechanics/search:
    get: #Search
      tags:
        - Mechanics
      summary: Search mechanics
      description: This endpoint allows you to search for mechanics by name and/or email.
        **At least one of `name` or `email` must be provided as a query parameter.**.
      parameters:
        - in: query
          name: name
          type: string
          description: Search term to filter mechanics by name.
        - in: query
          name: email
          type: string
          description: Search term to filter mechanics by email.
      responses:
        200:
          description: Return mechanics matching the search term
          schema:
            $ref: '#/definitions/MechanicListResponse'
          examples:
            application/json:
              [
                {
                  "id": 1,
                  "name": "John Doe",
                  "phone": "1234567890",
                  "email": "john@app.com",
                  "salary": 50000,
                  "service_tickets": [],
                }
              ]
        400:
          description: No query parameters provided
          schema:
            $ref: '#/definitions/MessageSchemaResponse'
          examples:
            application/json:
              message: At least one search parameter (name or email) is required.
              status: error
        404:
          description: No mechanics found
          schema:
            $ref: '#/definitions/MessageSchemaResponse'
          examples:
            application/json:
              message: No mechanics found.
              status: error

  # --------- Customers Endpoints ---------         
...
definitions: 
...
  LoginPayload:
    type: object
    properties:
      email:
        type: string
      password:
        type: string
  LoginResponse:
    type: object
    properties:
      status:
        type: string
      message:
        type: string
      token:
        type: string
  
  # --------- Message Schemas ---------
  MessageSchemaResponse:
    type: object
    properties:
      message:
        type: string
      status:
        type: string

  # --------- Mechanic Schemas ---------
  MechanicPayload:
    type: object
    properties:
      name:
        type: string
      phone:
        type: string
      email:
        type: string
      password:
        type: string
      salary:
        type: integer
  MechanicResponse:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      email:
        type: string
      salary:
        type: integer
      service_tickets:
        type: array     
  MechanicListResponse:
    type: array
    items:
      $ref: '#/definitions/MechanicResponse'
  
  #-------------- Customer Schemas-------
...